@@ -2,41 +2,44 @@ import 'dart:ui';
 
 import 'package:nav_router/nav_router.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:nav_router_example/pages/new_pahge.dart';
 
 class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key});
+
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  MyHomePageState createState() => MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
+class MyHomePageState extends State<MyHomePage> {
   List<RouteModel> data = [
-    new RouteModel('material', RouterType.material),
-    new RouteModel('cupertino', RouterType.cupertino),
-    new RouteModel('size', RouterType.size),
-    new RouteModel('fade', RouterType.fade),
-    new RouteModel('scale', RouterType.scale),
-    new RouteModel('scaleBottomRight', RouterType.scaleBottomRight),
-    new RouteModel('scaleTopLeft', RouterType.scaleTopLeft),
-    new RouteModel('slide', RouterType.slide),
-    new RouteModel('rotation', RouterType.rotation),
-    new RouteModel('scaleRotate', RouterType.scaleRotate),
-//    new RouteModel('routePushAdvance', RouterType.scaleRotate),
+    RouteModel('material', RouterType.material),
+    RouteModel('cupertino', RouterType.cupertino),
+    RouteModel('size', RouterType.size),
+    RouteModel('fade', RouterType.fade),
+    RouteModel('scale', RouterType.scale),
+    RouteModel('scaleBottomRight', RouterType.scaleBottomRight),
+    RouteModel('scaleTopLeft', RouterType.scaleTopLeft),
+    RouteModel('slide', RouterType.slide),
+    RouteModel('rotation', RouterType.rotation),
+    RouteModel('scaleRotate', RouterType.scaleRotate),
   ];
 
   Widget buildItem(RouteModel item) {
-    return new Container(
+    return SizedBox(
       width: (MediaQuery.of(context).size.width - 50) / 2,
-      child: new FlatButton(
-        color: Color(0xfff7c672),
-        padding: EdgeInsets.symmetric(vertical: 50.0),
-        onPressed: () => routePush(new NewPage(), item.type),
-//        onPressed: () => routePushAdvance(
-//            exitPage: new MyHomePage(), enterPage: new NewPage()),
-        child: new Text(
+      child: TextButton(
+        style: ButtonStyle(
+          padding: MaterialStateProperty.all(
+              const EdgeInsets.symmetric(vertical: 50.0)),
+
+          ///背景色
+          backgroundColor: MaterialStateProperty.all(const Color(0xfff7c672)),
+        ),
+        onPressed: () => routePush(const NewPage(), item.type),
+        child: Text(
           item.name,
-          style: TextStyle(color: Colors.black),
+          style: const TextStyle(color: Colors.black),
         ),
       ),
     );
@@ -44,23 +47,23 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
-    return new Scaffold(
-      backgroundColor: Color(0xfff4b050),
-      body: new Column(
+    return Scaffold(
+      backgroundColor: const Color(0xfff4b050),
+      body: Column(
         crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
-          new SizedBox(
-              height: MediaQueryData.fromWindow(window).padding.top * 2),
-          new Text(
+          SizedBox(height: MediaQueryData.fromWindow(window).padding.top * 2),
+          const Text(
             'NavRoute',
             style: TextStyle(color: Colors.blueAccent, fontSize: 30.0),
           ),
-          new SizedBox(height: 10),
-          new Expanded(
-            child: new SingleChildScrollView(
-              child: new Padding(
-                padding: EdgeInsets.symmetric(horizontal: 20, vertical: 15),
-                child: new Wrap(
+          const SizedBox(height: 10),
+          Expanded(
+            child: SingleChildScrollView(
+              child: Padding(
+                padding:
+                    const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+                child: Wrap(
                   spacing: 10.0,
                   runSpacing: 10.0,
                   alignment: WrapAlignment.start,
