@@ -7,7 +7,7 @@ class AdvanceSlideRoute extends PageRouteBuilder {
   final Widget enterPage;
   final Widget exitPage;
 
-  AdvanceSlideRoute({required this.exitPage,required  this.enterPage})
+  AdvanceSlideRoute({required this.exitPage, required this.enterPage})
       : super(
           pageBuilder: (
             BuildContext context,
@@ -24,14 +24,14 @@ class AdvanceSlideRoute extends PageRouteBuilder {
               Stack(
             children: <Widget>[
               SlideTransition(
-                position: new Tween<Offset>(
+                position: Tween<Offset>(
                   begin: const Offset(0.0, 0.0),
                   end: const Offset(-1.0, 0.0),
                 ).animate(animation),
                 child: exitPage,
               ),
               SlideTransition(
-                position: new Tween<Offset>(
+                position: Tween<Offset>(
                   begin: const Offset(1.0, 0.0),
                   end: Offset.zero,
                 ).animate(animation),
