@@ -2,46 +2,46 @@ import 'package:flutter/material.dart';
 
 /// [Route animation effect]
 /// Zoom and rotate mode route jump animation.
-/// 
+///
 class ScaleRotateRoute extends PageRouteBuilder {
   final Widget page;
   ScaleRotateRoute({required this.page})
       : super(
-    pageBuilder: (
-        BuildContext context,
-        Animation<double> animation,
-        Animation<double> secondaryAnimation,
-        ) =>
-    page,
-    transitionDuration: Duration(seconds: 1),
-    transitionsBuilder: (
-        BuildContext context,
-        Animation<double> animation,
-        Animation<double> secondaryAnimation,
-        Widget child,
-        ) =>
-        ScaleTransition(
-          scale: Tween<double>(
-            begin: 0.0,
-            end: 1.0,
-          ).animate(
-            CurvedAnimation(
-              parent: animation,
-              curve: Curves.fastOutSlowIn,
-            ),
-          ),
-          child: RotationTransition(
-            turns: Tween<double>(
+          pageBuilder: (
+            BuildContext context,
+            Animation<double> animation,
+            Animation<double> secondaryAnimation,
+          ) =>
+              page,
+          transitionDuration: const Duration(seconds: 1),
+          transitionsBuilder: (
+            BuildContext context,
+            Animation<double> animation,
+            Animation<double> secondaryAnimation,
+            Widget child,
+          ) =>
+              ScaleTransition(
+            scale: Tween<double>(
               begin: 0.0,
               end: 1.0,
             ).animate(
               CurvedAnimation(
                 parent: animation,
-                curve: Curves.linear,
+                curve: Curves.fastOutSlowIn,
               ),
             ),
-            child: child,
+            child: RotationTransition(
+              turns: Tween<double>(
+                begin: 0.0,
+                end: 1.0,
+              ).animate(
+                CurvedAnimation(
+                  parent: animation,
+                  curve: Curves.linear,
+                ),
+              ),
+              child: child,
+            ),
           ),
-        ),
-  );
+        );
 }
